+++ src/BrowserPicker.App/.editorconfig
