+++ src/BrowserPicker.App/App.config
