+++ src/BrowserPicker.App/App.xaml
