+++ src/BrowserPicker.App/App.xaml.cs
