+++ src/BrowserPicker.App/Converter/IconConverter.cs
