+++ src/BrowserPicker.App/Properties/AssemblyInfo.cs
@@ -47,4 +47,4 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.8.2.0")]
+[assembly: AssemblyVersion("1.9.0.0")]
