+++ src/BrowserPicker.App/Resources/privacy.png
