+++ src/BrowserPicker.App/Resources/web_icon.png
