+++ src/BrowserPicker.App/View/BrowserEditor.xaml
