+++ src/BrowserPicker.App/View/BrowserList.xaml.cs
