+++ src/BrowserPicker.App/View/Configuration.xaml
