+++ src/BrowserPicker.App/View/ExceptionReport.xaml.cs
