+++ src/BrowserPicker.App/View/LoadingWindow.xaml
