+++ src/BrowserPicker.App/View/LoadingWindow.xaml.cs
