+++ src/BrowserPicker.App/View/MainWindow.xaml
