+++ src/BrowserPicker.App/ViewModel/BrowserViewModel.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Windows;
-using BrowserPicker.Lib;
 using JetBrains.Annotations;
 
 namespace BrowserPicker.ViewModel
