+++ src/BrowserPicker.App/ViewModel/ExceptionViewModel.cs
@@ -1,5 +1,4 @@
-﻿using BrowserPicker.Lib;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
 using System;
 using System.Threading;
 using System.Windows;
