+++ src/BrowserPicker.App/packages.config
